Add null checks in HTML routes and title validation in PUT endpoint
- Add null guard in update route to handle non-existent todo_id
- Add null guard in delete route to handle non-existent todo_id
- Add title validation in API PUT endpoint to reject empty/null titles

+++ routes/todo.py
@@ -26,6 +26,8 @@ def add():
 @todo_bp.route("/update/<int:todo_id>")
 def update(todo_id):
     todo = Todo.query.filter_by(id=todo_id).first()
+    if not todo:
+        return redirect(url_for("todo.home"))
     todo.complete = not todo.complete
     db.session.commit()
     return redirect(url_for("todo.home"))
@@ -34,6 +36,8 @@ def update(todo_id):
 @todo_bp.route("/delete/<int:todo_id>")
 def delete(todo_id):
     todo = Todo.query.filter_by(id=todo_id).first()
+    if not todo:
+        return redirect(url_for("todo.home"))
     db.session.delete(todo)
     db.session.commit()
     return redirect(url_for("todo.home"))
